adding altered baseline, new model architecture file, and run model script
